Remove user management module and related templates, models, and routes
The user management functionality and associated templates, routes, and models have been fully removed from the codebase. This includes the dashboard, login, user editing, and profile handling features, as well as supporting models like User, Group, and Permission. This cleanup eliminates unused features and improves maintainability.

+++ librepos/manage.py
@@ -1,8 +1,6 @@
 import click
 
-from librepos.models import Group, Permission, Policy, Role, User, PermissionPolicy, PolicyGroup, GroupUser, MenuGroup, \
-    TicketType
-from .data.init_data import ROLES, PERMISSIONS, POLICIES, GROUPS, MENU_GROUPS, TICKET_TYPES
+from librepos.data.init_data import seed_all
 
 
 def add_cli_commands(app):
@@ -10,69 +8,22 @@ def add_cli_commands(app):
 
     from librepos.extensions import db
 
-    def populate_table(table, data):
-        for item in data:
-            table.create(**item)
+    @app.cli.command("dev", help="Development commands.")
+    @click.option("--reset-db", is_flag=True, help="Reset the database.")
+    @click.option("--seed-db", is_flag=True, help="Populate database with sample data.")
+    def dev(reset_db=False, seed_db=False):
+        """Development commands."""
+        if reset_db:
+            click.echo("Resetting the database...")
+            db.drop_all()
+            db.create_all()
+            click.echo("Done!")
 
-    @app.cli.command("initdb", help="Initialize the database.")
-    @click.option(
-        "--populate", is_flag=True, help="Populate database with sample data."
-    )
-    @click.confirmation_option(
-        prompt="This will drop existing data. Do you want to continue?",
-        help="Initialize the database.",
-    )
-    def init(populate):
-        """Initialize the database."""
-        click.echo("Initializing the database...")
-        db.drop_all()
-        db.create_all()
+        if seed_db:
+            click.echo("Populating database with sample data...")
 
-        populate_table(Permission, PERMISSIONS)
-        populate_table(Policy, POLICIES)
-        populate_table(Group, GROUPS)
-        populate_table(Role, ROLES)
-        populate_table(MenuGroup, MENU_GROUPS)
-        populate_table(TicketType, TICKET_TYPES)
+            seed_all()
 
-        # permission_policy
-        _permissions = Permission.query.all()
-        _administrator_access_policy = Policy.query.filter_by(
-            name="AdministratorAccess"
-        ).first()
+            click.echo("Done!")
 
-        if _administrator_access_policy:
-            for _permission in _permissions:
-                PermissionPolicy.create(
-                    permission_id=_permission.id,
-                    policy_id=_administrator_access_policy.id,
-                )
-
-        _administrator_group = Group.query.filter_by(name="Administrator").first()
-        if _administrator_group and _administrator_access_policy:
-            PolicyGroup.create(
-                group_id=_administrator_group.id,
-                policy_id=_administrator_access_policy.id,
-            )
-
-        # Create default admin
-        user_exists = User.query.count()
-        admin_role = Role.query.filter_by(name="admin").first()
-
-        if not user_exists and admin_role:
-            User.create(
-                role_id=admin_role.id,
-                password="librepos",
-                username="LibrePOS",
-            )
-
-        # add the newly create admin user to the Administrator Group
-        admin_group = Group.query.filter_by(name="Administrator").first()
-        if admin_group:
-            admin_user = User.query.filter_by(username="LibrePOS").first()
-            if admin_user:
-                GroupUser.create(group_id=admin_group.id, user_id=admin_user.id)
-
-        click.echo("Done!")
-
-    app.cli.add_command(init)
+    app.cli.add_command(dev)
